requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
pytesseract
PyMuPDF
transformers
torch
requests
